# Fix vtproto pooling to reduce allocs