@dmathieu 2023-07-11
That seems like something which can break quite easily. Would there be a way to do it in new methods?