@carsonip 2023-07-11
> That seems like something which can break quite easily. Would there be a way to do it in new methods?

@dmathieu I agree. Created #17 to fix codegen (it also fixes another instance of ineffective pooling). We'll discuss whether to maintain a fork or to fix upstream. Closing this PR in favor of #17